Fix warnings

+++ src/xrGameSpy/gamespy/GP/gpiPeer.c
@@ -272,7 +272,7 @@ gpiProcessPeerAcceptingConnection(
 				iconnection->password,
 				iconnection->profileid,
 				pid);
-			MD5Digest((unsigned char *)buffer, strlen(buffer), sigCheck);
+			MD5Digest((unsigned char *)buffer, (unsigned int)strlen(buffer), sigCheck);
 
 			// Check the sig.
 			/////////////////
